ts rewrite
